+++ src/composables/index.ts
@@ -1,0 +1,5 @@
+export { useCities } from './useCities'
+export type { CityData } from './useCities'
+
+export { useHistory } from './useHistory'
+export type { HistoryData, HistoryDataPoint, HistoryRange } from './useHistory'